Fix segmentation fault using project in fill_with_density2
When using xyhalfshift2reg to get the region id in function project for
applying a different scaling of the displacement vector all iterations >
1 resulted in a segmentation fault. Probably this has to do with
allocating and freeing memory for xyhalfshift2reg in fill_with_density1
and 2 thus resulting in invalid memory access in project. A workaround
for now is to allocate memory once in fill_with_density1 and free it at
the end of the whole procedure.